Channel and layout

+++ backend/src/channel-member/channel-member.resolver.ts
@@ -1,4 +1,11 @@
-import { Args, Mutation, Query, Resolver, Subscription } from '@nestjs/graphql'
+import {
+  Args,
+  Context,
+  Mutation,
+  Query,
+  Resolver,
+  Subscription
+} from '@nestjs/graphql'
 import { ChannelMember } from './entities/channel-member.entity'
 import { ChannelMemberService } from './channel-member.service'
 import { CreateChannelMemberInput } from './dto/create-channel-member.input'
@@ -10,13 +17,26 @@ import {
   Unprotected
 } from '../auth/guards/authorization.guard'
 import { PubSub } from 'graphql-subscriptions'
+import {
+  ForbiddenAccessChannelAdmin,
+  ForbiddenAccessChannelOwner,
+  ForbiddenAccessData,
+  channelAdminGuard,
+  channelOwnerGuard,
+  userContextGuard
+} from 'src/auth/guards/request.guards'
+import { PrismaService } from 'src/prisma/prisma.service'
+import { ExceptionTryingToDeleteChannelOwner } from 'src/channel/exceptions/channel-member.exceptions'
+import * as bcrypt from 'bcrypt'
+import { ExceptionWrongChannelPassword } from 'src/channel/exceptions/channel.exception'
 
 @Resolver(() => ChannelMember)
 @UseGuards(AuthorizationGuard)
 export class ChannelMemberResolver {
   constructor(
     private readonly channelMemberService: ChannelMemberService,
-    private readonly pubSub: PubSub
+    private readonly pubSub: PubSub,
+    private readonly prisma: PrismaService
   ) {}
 
   //**************************************************//
@@ -64,8 +84,22 @@ export class ChannelMemberResolver {
   @Mutation(() => ChannelMember)
   async createChannelMember(
     @Args('data', { type: () => CreateChannelMemberInput }, ValidationPipe)
-    data: CreateChannelMemberInput
+    data: CreateChannelMemberInput,
+    @Context() ctx: any
   ): Promise<ChannelMember> {
+    if (!userContextGuard(ctx?.req?.user?.id, data.userId))
+      throw new ForbiddenAccessData()
+
+    const channel = await this.prisma.channel.findUnique({
+      where: { id: data.channelId }
+    })
+
+    if (
+      channel?.password &&
+      !(await bcrypt.compare(channel.password, data.channelPassword as string))
+    )
+      throw new ExceptionWrongChannelPassword()
+
     const res = await this.channelMemberService.create(data)
 
     const channelMembers = await this.findAllChannelMemberInChannel(
@@ -76,7 +110,9 @@ export class ChannelMemberResolver {
         await this.pubSub.publish('memberReceived-' + value.userId, { res })
       })
     )
-    await this.pubSub.publish('memberReceived-' + data.channelId, { res })
+
+    await this.pubSub.publish('memberReceived-' + res.userId, { res })
+
     return res
   }
 
@@ -87,11 +123,20 @@ export class ChannelMemberResolver {
     @Args('channelId', { type: () => String }, NanoidValidationPipe)
     channelId: string,
     @Args('data', { type: () => UpdateChannelMemberInput }, ValidationPipe)
-    data: UpdateChannelMemberInput
+    data: UpdateChannelMemberInput,
+    @Context() ctx: any
   ): Promise<ChannelMember> {
+    if (!userContextGuard(ctx?.req?.user?.id, userId))
+      throw new ForbiddenAccessData()
     const res = await this.channelMemberService.update(userId, channelId, data)
 
-    await this.pubSub.publish('memberReceived-' + channelId, { res })
+    const channelMembers = await this.findAllChannelMemberInChannel(channelId)
+    await Promise.all(
+      channelMembers.map(async (value) => {
+        await this.pubSub.publish('memberEdited-' + value.userId, { res })
+      })
+    )
+
     return res
   }
 
@@ -100,11 +145,32 @@ export class ChannelMemberResolver {
     @Args('userId', { type: () => String }, NanoidValidationPipe)
     userId: string,
     @Args('channelId', { type: () => String }, NanoidValidationPipe)
-    channelId: string
+    channelId: string,
+    @Context() ctx: any
   ): Promise<ChannelMember> {
+    if (
+      !(await channelAdminGuard(ctx?.req?.user?.id, channelId, this.prisma)) &&
+      !userContextGuard(ctx?.req?.user?.id, userId)
+    )
+      throw new ForbiddenAccessData()
+
+    const channel = await this.prisma.channel.findUnique({
+      where: { id: channelId }
+    })
+
+    if (!channel || channel.ownerId === userId)
+      throw new ExceptionTryingToDeleteChannelOwner()
+
     const res = await this.channelMemberService.delete(userId, channelId)
 
-    await this.pubSub.publish('memberReceived-' + channelId, { res })
+    const channelMembers = await this.findAllChannelMemberInChannel(channelId)
+    await Promise.all(
+      channelMembers.map(async (value) => {
+        await this.pubSub.publish('memberDeleted-' + value.userId, { res })
+      })
+    )
+
+    await this.pubSub.publish('memberDeleted-' + userId, { res })
     return res
   }
 
@@ -113,9 +179,22 @@ export class ChannelMemberResolver {
     @Args('userId', { type: () => String }, NanoidValidationPipe)
     userId: string,
     @Args('channelId', { type: () => String }, NanoidValidationPipe)
-    channelId: string
+    channelId: string,
+    @Context() ctx: any
   ): Promise<ChannelMember> {
-    return this.channelMemberService.unmakeAdmin(userId, channelId)
+    if (!(await channelOwnerGuard(ctx?.req?.user?.id, channelId, this.prisma)))
+      throw new ForbiddenAccessChannelOwner()
+    const res = await this.channelMemberService.unmakeAdmin(userId, channelId)
+
+    const channelMembers = await this.findAllChannelMemberInChannel(channelId)
+    await Promise.all(
+      channelMembers.map(async (value) => {
+        await this.pubSub.publish('memberEdited-' + value.userId, { res })
+      })
+    )
+
+    await this.pubSub.publish('memberEdited-' + channelId, { res })
+    return res
   }
 
   @Mutation(() => ChannelMember)
@@ -123,9 +202,22 @@ export class ChannelMemberResolver {
     @Args('userId', { type: () => String }, NanoidValidationPipe)
     userId: string,
     @Args('channelId', { type: () => String }, NanoidValidationPipe)
-    channelId: string
+    channelId: string,
+    @Context() ctx: any
   ): Promise<ChannelMember> {
-    return this.channelMemberService.makeAdmin(userId, channelId)
+    if (!(await channelOwnerGuard(ctx?.req?.user?.id, channelId, this.prisma)))
+      throw new ForbiddenAccessChannelOwner()
+    const res = await this.channelMemberService.makeAdmin(userId, channelId)
+
+    const channelMembers = await this.findAllChannelMemberInChannel(channelId)
+    await Promise.all(
+      channelMembers.map(async (value) => {
+        await this.pubSub.publish('memberEdited-' + value.userId, { res })
+      })
+    )
+
+    await this.pubSub.publish('memberEdited-' + channelId, { res })
+    return res
   }
 
   @Mutation(() => ChannelMember)
@@ -133,9 +225,22 @@ export class ChannelMemberResolver {
     @Args('userId', { type: () => String }, NanoidValidationPipe)
     userId: string,
     @Args('channelId', { type: () => String }, NanoidValidationPipe)
-    channelId: string
+    channelId: string,
+    @Context() ctx: any
   ): Promise<ChannelMember> {
-    return this.channelMemberService.mute(userId, channelId)
+    if (!(await channelAdminGuard(ctx?.req?.user?.id, channelId, this.prisma)))
+      throw new ForbiddenAccessChannelAdmin()
+    const res = await this.channelMemberService.mute(userId, channelId)
+
+    const channelMembers = await this.findAllChannelMemberInChannel(channelId)
+    await Promise.all(
+      channelMembers.map(async (value) => {
+        await this.pubSub.publish('memberEdited-' + value.userId, { res })
+      })
+    )
+
+    await this.pubSub.publish('memberEdited-' + channelId, { res })
+    return res
   }
 
   @Mutation(() => ChannelMember)
@@ -143,9 +248,22 @@ export class ChannelMemberResolver {
     @Args('userId', { type: () => String }, NanoidValidationPipe)
     userId: string,
     @Args('channelId', { type: () => String }, NanoidValidationPipe)
-    channelId: string
+    channelId: string,
+    @Context() ctx: any
   ): Promise<ChannelMember> {
-    return this.channelMemberService.unmute(userId, channelId)
+    if (!(await channelAdminGuard(ctx?.req?.user?.id, channelId, this.prisma)))
+      throw new ForbiddenAccessChannelAdmin()
+    const res = await this.channelMemberService.unmute(userId, channelId)
+
+    const channelMembers = await this.findAllChannelMemberInChannel(channelId)
+    await Promise.all(
+      channelMembers.map(async (value) => {
+        await this.pubSub.publish('memberEdited-' + value.userId, { res })
+      })
+    )
+
+    await this.pubSub.publish('memberEdited-' + channelId, { res })
+    return res
   }
 
   //**************************************************//
